Move some test utils from test to testing
This fixes the Docker build. "test" was added to .dockerignore.
