+++ cmd/sgnops.go
@@ -4,8 +4,8 @@ import (
 	"github.com/celer-network/sgn/app"
 	"github.com/celer-network/sgn/common"
 	"github.com/celer-network/sgn/ops"
-	"github.com/celer-network/sgn/test/channel"
-	tc "github.com/celer-network/sgn/test/common"
+	"github.com/celer-network/sgn/testing/channel"
+	tc "github.com/celer-network/sgn/testing/common"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"github.com/tendermint/tendermint/libs/cli"
