+++ test/e2e/multinode/e2e_setup_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/celer-network/goutils/log"
 	"github.com/celer-network/sgn/mainchain"
-	tc "github.com/celer-network/sgn/test/common"
+	tc "github.com/celer-network/sgn/testing/common"
 )
 
 // TestMain handles common setup (start mainchain, deploy, start sidechain etc)
