+++ test/e2e/multinode/setup_sgn.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/celer-network/goutils/log"
 	"github.com/celer-network/sgn/common"
-	tc "github.com/celer-network/sgn/test/common"
+	tc "github.com/celer-network/sgn/testing/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/spf13/viper"
 )
