+++ test/e2e/singlenode/e2e_setup_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/celer-network/goutils/log"
 	"github.com/celer-network/sgn/mainchain"
-	tc "github.com/celer-network/sgn/test/common"
+	tc "github.com/celer-network/sgn/testing/common"
 )
 
 var (
