+++ test/e2e/singlenode/setup_sgn.go
@@ -1,5 +1,3 @@
-// Copyright 2018 Celer Network
-
 package singlenode
 
 import (
@@ -11,7 +9,7 @@ import (
 
 	"github.com/celer-network/goutils/log"
 	"github.com/celer-network/sgn/common"
-	tc "github.com/celer-network/sgn/test/common"
+	tc "github.com/celer-network/sgn/testing/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/spf13/viper"
 )
