+++ testing/channel-eth-go/deploy/deploy.go
@@ -10,18 +10,18 @@ import (
 	"strings"
 	"time"
 
-	"github.com/celer-network/sgn/test/channel-eth-go/balancelimit"
-	"github.com/celer-network/sgn/test/channel-eth-go/channel"
-	"github.com/celer-network/sgn/test/channel-eth-go/ethpool"
-	"github.com/celer-network/sgn/test/channel-eth-go/ledger"
-	"github.com/celer-network/sgn/test/channel-eth-go/ledgerstruct"
-	"github.com/celer-network/sgn/test/channel-eth-go/migrate"
-	"github.com/celer-network/sgn/test/channel-eth-go/operation"
-	"github.com/celer-network/sgn/test/channel-eth-go/payregistry"
-	"github.com/celer-network/sgn/test/channel-eth-go/payresolver"
-	"github.com/celer-network/sgn/test/channel-eth-go/routerregistry"
-	"github.com/celer-network/sgn/test/channel-eth-go/virtresolver"
-	"github.com/celer-network/sgn/test/channel-eth-go/wallet"
+	"github.com/celer-network/sgn/testing/channel-eth-go/balancelimit"
+	"github.com/celer-network/sgn/testing/channel-eth-go/channel"
+	"github.com/celer-network/sgn/testing/channel-eth-go/ethpool"
+	"github.com/celer-network/sgn/testing/channel-eth-go/ledger"
+	"github.com/celer-network/sgn/testing/channel-eth-go/ledgerstruct"
+	"github.com/celer-network/sgn/testing/channel-eth-go/migrate"
+	"github.com/celer-network/sgn/testing/channel-eth-go/operation"
+	"github.com/celer-network/sgn/testing/channel-eth-go/payregistry"
+	"github.com/celer-network/sgn/testing/channel-eth-go/payresolver"
+	"github.com/celer-network/sgn/testing/channel-eth-go/routerregistry"
+	"github.com/celer-network/sgn/testing/channel-eth-go/virtresolver"
+	"github.com/celer-network/sgn/testing/channel-eth-go/wallet"
 	ethereum "github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
