+++ testing/channel/channel.go
@@ -15,7 +15,7 @@ import (
 	"github.com/celer-network/sgn/app"
 	"github.com/celer-network/sgn/common"
 	"github.com/celer-network/sgn/mainchain"
-	tc "github.com/celer-network/sgn/test/common"
+	tc "github.com/celer-network/sgn/testing/common"
 	"github.com/celer-network/sgn/transactor"
 	"github.com/celer-network/sgn/x/guard"
 	"github.com/celer-network/sgn/x/sync"
