+++ testing/common/constants.go
@@ -1,4 +1,4 @@
-package testcommon
+package common
 
 import "time"
