+++ testing/common/framework.go
@@ -1,5 +1,4 @@
-// Copyright 2018 Celer Network
-package testcommon
+package common
 
 import (
 	"os"
