+++ testing/common/sidechain.go
@@ -1,4 +1,4 @@
-package testcommon
+package common
 
 import (
 	"errors"
